First version draft of how State Machine execution will look like
Created the main classes of any State Machine: States and Transitions.
States map to transisitions through their index, and transitions have the main elements: events, guards and actions

The State Machine includes more than states and transitions, it has an environment to match the boolean expressions of the guards against it, and a set of events that are to be checked at each run of the State Machine.

Also included is an example of how to create a State Machine with 3 states and 4 transitions.