Added javadoc documentation and defined a end State signal
Most of the class now include Javadoc comments

An end State is a state with no transitions, in which case the State Machine can not perform any more actions and has reached a stationary state. This signal is now returned by the update() method that is invoke en each execution of the State Machine.