Simplied the Event class and added logic operators
The Event class now represents only one event, and not a the set of all
Events that the Transisitons have. To work in this way further changes
to the example of State Machine are necessary.

Added a class to represent logic operators, that can be used by events
or guards. Before the relation was an implicit AND, all the events and
guards had to be true for the transitions to be activated/triggered.
Now we define 4 operators: NOT, AND, OR and XOR. To use them we need to
have combinations of one operand, for NOT, of two operands, for XOR,
and of two or more operands for AND and OR.