Finished LogicOperator class with custom exception
The LogicOperator class supports NOT, OR, AND and XOR operations. If the
number of operands for the logic operator is not correct, it will throw
a custom checked exception with the message explaining the problem.

The custom checked exception is called IllegalNumberOfOperandsException.
For each case it will add an error message to the constructor specific
to that operator.