Added LogicExpression to group Operator and Operands
The new class LogicExpression puts together an Operator with a list of
Operands, such that it can be apply the same operator to all the
operands. Now the Transition class instead of only containing an Event
and a Guard, it can contain multiple Events or Guards, and the relation
between the operands can be any of the Operators supported.

For this purpose I have also added an interface Evaluable, implemented
by those classes that can be contained in a LogicExpression.

The javadoc documentation for the new classes has also been added.