Started to update State Machine example in main
The example in main was outdated since all the changes done in previous
commits. I have created a similar State Machine as in previous example,
and tested limit cases of Transitions, when there is no Event, no Guard
or no Action. Also tested the logic expressions and logic operators
with these examples.

Results showed many problems, mainly NullPointer exceptions. From these
problems I stablished that the State Machine accepts transitions
without Guards or Events, and added a TODO comment at these lines for
finding the lines easier in future times.

Also found problems when creating State Machines with incorrect States
or Transitions, that point to nonexistent objects. Added test method
to find these problems when creating the State Machine object.

Another problem was found when evaluating the Events, the method was
outdated from where the Events where a Set of Strings, and not a Set of
Objects. Needed to rewrite the evaluate method to adapt to the change.

The last problem was for logical expressions without operator, and only
one operand. In these case we only evaluate the operand, but need to
account for incorrect inputs. For that case I created another checked
Exception, IllegalLogicExpressionException.

Also did some renaming and addition of getters and setters method when
necessary.