Started unit testing Event after main example work
The example of State Machine is finally working with the new additions
of logic expressions and operators, further test will be conducted in
the future.

Wrote the first unit tests for the Event class to see more in depth the
limitations of the class and those border cases that I could not find
in the general example.

Also fixed some typos