Started work on Actions (Inheritance or Functional)
Previously created a class Action, and used inheritance to change its
behaviour by extending it(subclass) and overriding the trigger method.
To the initial class I have added as input the Events, so that the
class has all the context needed for its business logic.

Now I created a different class, as another alternative to the original
Action class, that uses a functinal interface as a way to change its
behaviour. In this design, to create a custom action only requires to
pass as argument to the constructor the method that the action must
implement, rather than creating a whole new class and overriding the
parent class method.