Changed Action class to functional interface #1
After proposing design changes to Action class, the conclusion was that
it would be a functional interface itself. There is no need for it to
have a field of type functional interface, is simpler this way.

There was no argument against implementing it with lambda expressions,
as the use case where it would need an object context is not in scope at
the moment.

This also simplifies action creation, as will be demonstrated in the
main class, in the State Machine example.