Fixed bug with Events being cleared before used
The bug appeared when a transition issued and Event, it would have been
included in the Events Set for a brief period of time, and removed as
soon as all the transitions were checked. That meant the Event would not
be active the next time the State Machine checks the transitions, which
was an error.

Solution works under the assumption that only one transitions can be
activated in each State. For parallel States and multiple transitions
this solution will need to be modified.