Change StateMachine and update Main accordingly
Fix format of StateMachine code, add more get and set methods for
better usability, add copy constructor and change return type of update
method.

The copy constructor enables to have State Machines as templates, and
made a copy for execution without modifying the original.

There was a confusion in getCurrentState() method, as name was used in
the class to refer to the field storing the number of the state, but
the method returned the state object. Created separate methods, one to
retrieve the number and another for the object, and added a set method
to introduce the number of the State that will be active and used to
update the StateMachine. This is useful, for example, to reset the
State Machine.

Changed the return value of update(). Before, the returned value was a
bool, signaling with a false value the end state of the State Machine.
Now it returns a wrapper object that contains more information, like
how was the update of the State Machine (was a transition triggered?),
in addition to also signaling the end state. The object also includes a
reference to the state and transition, this way the user of the State
Machine can know which of the many transitions was taken to arrive to
that state.